[#4] fix : review Fix
추가사항 : scss 압축용 플러그인 추가
