+++ client/package.json
@@ -24,6 +24,7 @@
     "html-webpack-plugin": "^5.3.2",
     "mini-css-extract-plugin": "^2.1.0",
     "node-sass": "^6.0.1",
+    "optimize-css-assets-webpack-plugin": "^6.0.1",
     "sass-loader": "^12.1.0",
     "ts-loader": "^9.2.3",
     "ts-node": "^10.1.0",
