+++ client/src/components/header/index.scss
@@ -15,7 +15,7 @@
 
     @media (max-width: $table-max-width) {
       grid-template-columns: 1fr;
-      grid-row-gap: .9em;
+      grid-row-gap: 0.9em;
     }
 
     li.left {
@@ -48,26 +48,27 @@
         }
       }
     }
-    li.right {
-      justify-self: end;
-      font-size: 1.3em;
-      display: flex;
-      justify-content:space-between;
-      gap: .5em;
-      div {
-        svg {
-          width: 1.8em;
-          height: 1.8em;
-          path {
-            stroke: $off-white;
-          }
-        }
-        &.selected {
-          svg {
-            path {
-              stroke-width: 2px;
-            }
-          }
+  }
+}
+
+li.header-wrap-right {
+  justify-self: end;
+  font-size: 1.3em;
+  display: flex;
+  justify-content: space-between;
+  gap: 0.5em;
+  .svg-icon {
+    svg {
+      width: 1.8em;
+      height: 1.8em;
+      path {
+        stroke: $off-white;
+      }
+    }
+    &.selected {
+      svg {
+        path {
+          stroke-width: 2px;
         }
       }
     }
